[v0.22.1] [Refactor] Extract assistant core into shared library
