+++ src/desktop/app/XerahS.Assistant/XerahS.Assistant.csproj
@@ -1,0 +1,29 @@
+<Project Sdk="Microsoft.NET.Sdk">
+  <PropertyGroup>
+    <RootNamespace>XerahS.Assistant</RootNamespace>
+    <IsWindowsAssistantBuild>false</IsWindowsAssistantBuild>
+    <IsMacOSAssistantBuild>false</IsMacOSAssistantBuild>
+    <IsLinuxAssistantBuild>false</IsLinuxAssistantBuild>
+    <IsWindowsAssistantBuild Condition="$(RuntimeIdentifier.StartsWith('win')) Or ('$(OS)' == 'Windows_NT' And '$(RuntimeIdentifier)' == '')">true</IsWindowsAssistantBuild>
+    <IsMacOSAssistantBuild Condition="$(RuntimeIdentifier.StartsWith('osx')) Or ('$(RuntimeIdentifier)' == '' And $([MSBuild]::IsOSPlatform('OSX')))">true</IsMacOSAssistantBuild>
+    <IsLinuxAssistantBuild Condition="$(RuntimeIdentifier.StartsWith('linux')) Or ('$(RuntimeIdentifier)' == '' And $([MSBuild]::IsOSPlatform('Linux')))">true</IsLinuxAssistantBuild>
+    <TargetFramework>net10.0</TargetFramework>
+    <TargetFramework Condition="'$(IsWindowsAssistantBuild)' == 'true'">net10.0-windows10.0.26100.0</TargetFramework>
+    <Nullable>enable</Nullable>
+    <ImplicitUsings>enable</ImplicitUsings>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.Data.Sqlite" />
+    <PackageReference Include="Microsoft.Extensions.DependencyInjection" />
+    <PackageReference Include="SkiaSharp" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\core\XerahS.Common\XerahS.Common.csproj" />
+    <ProjectReference Include="..\..\core\XerahS.Core\XerahS.Core.csproj" />
+    <ProjectReference Include="..\..\core\XerahS.History\XerahS.History.csproj" />
+    <ProjectReference Include="..\XerahS.Bootstrap\XerahS.Bootstrap.csproj" />
+    <ProjectReference Include="..\..\..\platform\XerahS.Platform.Abstractions\XerahS.Platform.Abstractions.csproj" />
+  </ItemGroup>
+</Project>
